requirements.txt
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
backcall==0.1.0
certifi==2018.10.15
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
daphne==2.2.3
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
future==0.17.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==7.1.1
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.13.1
Markdown==3.0.1
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.0
PyHamcrest==1.9.0
pyserial==3.4
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
stripe==2.12.0
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
wcwidth==0.1.7
zope.interface==4.6.0
